starting ui

+++ lib/movies/data/models/movie_model.dart
@@ -18,7 +18,7 @@ class MovieModel extends Movie {
       backdropPath: json["backdrop_path"],
       genreIds: List<int>.from(json["genre_ids"].map((e) => e)),
       overview: json["overview"],
-      voteAverage: json["vote_average"],
+      voteAverage: json["vote_average"].toDouble(),
       releaseDate: json["release_date"],
     );
   }
